+++ R/read.r
@@ -111,10 +111,16 @@ read_ebd.character <- function(x, reader, sep = "\t", unique = TRUE,
 
   # remove duplicate group checklists
   if (unique) {
+    if (!"group_identifier" %in% names(out)) {
+      stop("unique = TRUE used but group_identifier not found in data.")
+    }
     out <- auk_unique(out)
   }
   # taxonomic rollup
   if (rollup) {
+  if (!"scientific_name" %in% names(out)) {
+    stop("rollup = TRUE used but scientific_name not found in data.")
+  }
     out <- auk_rollup(out)
   }
   row.names(out) <- NULL